# Optimise ImageOps.fit by combining resize and crop