--- a/docs/releasenotes/5.4.0.rst
+++ b/docs/releasenotes/5.4.0.rst
@@ -8,7 +8,7 @@ Other Changes
 ImageOps.fit
 ^^^^^^^^^^^^
 
-Now uses the one resize operation with ``box`` argument internally
-instead of the crop and scale operations sequence.
+Now uses one resize operation with ``box`` parameter internally
+instead of a crop and scale operations sequence.
 This improves the performance and accuracy of cropping since
-``box`` parameter accepts float values.
+the ``box`` parameter accepts float values.