@homm 2018-10-18
@hugovk Added release notes. Could you check?